added linaer stt unit tests
