+++ tests/test_miramind/test_audio/test_stt/test_stt_class.py
@@ -1,4 +1,8 @@
-from miramind.audio.stt.stt_class import STT
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+from miramind.audio.stt.stt_class import STT, LinearListeningSTT
 
 
 def test_STT(mocker):
@@ -10,3 +14,17 @@ def test_STT(mocker):
     mocker.patch("builtins.open")
     stt = STT(client=mock_azure_openai_client)
     assert stt.transcribe("fake file") == {"transcript": "fake transcript"}
+
+
+def test_linear_stt(mocker):
+    mocker.patch("miramind.audio.stt.stt_class.sd.rec", return_value=None)
+    mocker.patch("miramind.audio.stt.stt_class.sd.wait")
+    mocker.patch("miramind.audio.stt.stt_class.sf.write")
+    mock_transcript = mocker.Mock()
+    mock_transcript.text = "fake transcript"
+    mock_azure_openai_client = mocker.MagicMock()
+    mock_azure_openai_client.audio.transcriptions.create.return_value = mock_transcript
+
+    llstt = LinearListeningSTT(client=mock_azure_openai_client)
+    result = llstt.run()
+    assert result["transcript"] == "fake transcript"
